requirements.txt
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
attrs==25.3.0
certifi==2025.4.26
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.0
croniter==6.0.0
cryptography==44.0.3
cssselect2==0.8.0
defusedxml==0.7.1
Django==5.1.7
django-simple-history==3.8.0
django-tenants==3.7.0
django-unfold==0.57.0
html5lib==1.1
idna==3.10
lxml==5.4.0
numpy==2.2.5
oscrypto==1.3.0
pillow==11.2.1
procrastinate==3.1.0
psycopg==3.2.7
psycopg-binary==3.2.7
psycopg-pool==3.2.6
pycparser==2.22
pyHanko==0.26.0
pyhanko-certvalidator==0.26.8
pypdf==5.5.0
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.2
qrcode==8.2
reportlab==4.4.0
requests==2.32.3
setuptools==80.4.0
six==1.17.0
sqlparse==0.5.3
svglib==1.5.1
tinycss2==1.4.0
typing_extensions==4.13.2
tzlocal==5.3.1
uritools==5.0.0
urllib3==2.4.0
webencodings==0.5.1
wheel==0.45.1
xhtml2pdf==0.2.17
